services/inventory-service/requirements.txt
grpcio
grpcio-tools
protobuf
requests